Cave Filtering
Added function to get a list of tiles in the form of tuples (x, y) that correspond to all the separate caves in a given tilemap.

Then used this to filter out caves below a size of 8 and also used it to get an unoccupied tile in which to spawn the player on.

May be an issue in the future with randomness, the level can have a seed will be different to that in the game class, may cause an issue where the same seed will spawn you in different places which isn't necessarily bad.