Refactor of components parent entity assignment
Now assignment of the parent entity of a component is done in the add_component method of the entity class. This means that you no longer have to specify a parent in the init function of each component which saves a lot of writing. To fix further issues an on_add_component method can be added to any component and will be called after a parent entity is assigned in add_component on the entity class