Added ItemActions and Entity Item Rework
ItemActions for items so that they can be used in different ways.
Replaced Energizium Item Entity with standard item entity that contains the item that is inside it.
This rework means the character and foreground colour of the entity is stored on the item
This rework also allowed for the item graphic to be shown in the inventory ui