Equipment and Equpiment Items
EQUIPMENT AND EQUIPMENT ITEMS
Added Equipment component for keeping track of what items are equipped on an entity
Added EquipmentItem subclass to the Item class which allows for items to be equipped
Added basic items for equipping in each slot
Added visual representation of equipped items to the inventory_renderer
Changed Input_Handler so that the action of the item in the left or right hand of the player is used. This means you can only mine with the hand with the laser cutter, shoot with the hand with the gun ect.

OTHER CHANGES
Reworked items to always have the drop action, don't need to add it in the parameters now
Added drop pod entity and made it randomly spawn in the starting cave. Currently doesn't do anything but later will be used for crafting/research?
Added damage utility class to later be used for calculating damages